Spread out ProducerStage contents over multiple files
